ASACORE-2030 Validate public keys at import
Validate that the public key is a point on the elliptic curve when
importing, instead of later when the key is first used.

Move away from memcpy_s in Import in the CNG implementation; any
failures here would cause the object to be left in an inconsistent
state. Add some assertions instead to check for the overrun
conditions memcpy_s looks for.

Correct some comments in the CryptoECC.h header file.

Add a unit test to take a valid key pair, randomly generate a new Y
coordinate, and verify that it doesn't import successfully and the
ECCPublicKey remains unchanged.

Change-Id: Iec3217fb03913a52c4cb4a025c1ca5973b4269c1
Signed-off-by: Kevin Kane <kkane@microsoft.com>