ASACORE-2644 Multiple tests in AJTest are running >800 times slower on Windows 10
A recent change (ASACORE-1317, commit fc4818082f8e516a28d1c5df2f777b764e39ca04)
causes multiple tests in AJTest suite to appear hung for at least 30 seconds
during the teardown process (on Windows 10). This is due to the RemoteEndpoint
change that introduces a 30-second delay (by default) to wait for the peer to
respond while the teardown routine calls Stop() and Join().

The fix here is to change ClientTransport::Disconnect (a super class of
NamedPipeClientTransport) to call StopAfterTxEmpty. Doing so forces an exit
callback on the RemoteEndpoint due to the stream closure (when nothing to TX),
effectively makes Join return much faster.

Note that the new behavior introduced by the above commit makes the default
Join() wait 30 seconds for the remote peer to respond, which is not a
requirement for named pipe since the data is already sitting in the kernel
buffer (thus delivery is guaranteed as long as the peer is up).

Change-Id: Iccce6d537aa3424361b568a11cc1e4bd00507717
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>