Additional documentation
This add an image that shows many of the components of
a BusAttachment. This image was created when teaching
other developers about AllJoyn.

The feedback that I had was that the image was useful
as a reference when trying to first build an AllJoyn
application.  For this reason I am including the document
image in the document.

Change-Id: Ic9406c6e149cad4ad0d8bef4d7f3c0b3e43b5683
Signed-off-by: Todd Malsbary <tmalsbar@qce.qualcomm.com>