ASACORE-2660 Enforce ClaimCapabilities when an app is claimed
ClaimCapabilities are enforced when the Claim method is called,
bringing the SCL's behavior in line with what is currently
implemented in the TCL.

To avoid breaking existing apps, default the ClaimCapabilities to
be both ECDHE_NULL and ECDHE_PSK. The app still has to configure
security on its bus attachment to accept the desired key exchange
method.

Miscellaneous test fixes to expect this new default value, and some
spelling fixes en route.

Unrelated, check the return value of a couple calls to
ECCPublicKey::Import in KeyExchanger that weren't being checked.

Change-Id: Id86b4273431d3ce80bdb481ea8dc88177e566edc
Signed-off-by: Kevin Kane <kkane@microsoft.com>