ASACORE-2699 Improve efficiency of qcc::GetUid and qcc::GetGid on Windows
GetUid and GetGid currently account to 80% of the delay in processing DBus's
NameOwnerChanged signals (as well as other common functions such as
BusAttachment::Connect). A NameOwnerChanged (sessionless) signal is generated
each time a BusAttachment's connection state changes. Since AllJoyn Routers
broadcast these signals to everyone on the network, reducing delays in this
code path is important on large networks. On Windows, qcc::GetUid and
qcc::GetGid call the OS API GetUserNameExA (sspicli.dll), which is a relatively
expensive call.

This change moves this OS call to the initialization phase, effectively making
it one-time cost.

Change-Id: I810c58d10a0b52999fd7c235485eb05bab991dc4
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>