ASACORE-2696 Improve Router CPU utilization when dealing with About signals
As the number of AllJoyn BusAttachments in the network increases, the CPU
utilization in AllJoyn (router) gets worse. The performance bottleneck is
especially high when the router needs to match who-implements rules against
the About sessionless signals in the network. Profiler runs against a
network of 200 Producer BusAttachments found this bottleneck to contribute
more than half of the overall delay during discovery and session
establishment, particularly by the following callstack:

    ...
    ajn::_Message::UnmarshalArgs
    ajn::Rule::IsMatch
    ajn::SessionlessObj::IsOnlyImplicitMatch
    ajn::SessionlessObj::SendMatchingThroughEndpoint
    ajn::SessionlessObj::HandleRangeRequest
    ajn::SessionlessObj::AddRuleWork::Run
    ajn::SessionlessObj::AlarmTriggered

Because the routine _Message::UnmarshallArgs is expensive, reducing the number
of calls is important for scalability. This fix is to limit the frequency of
having to unnecessarily parse the same About signal over and over by caching
the who-implements results (interfaces). This allows rule matching routine
[Rule::IsMatch] to become more efficient and thus reduces the number of calls
to UnmarshalArgs.

Change-Id: I2b14277b3aabc3eb464c6217cc756e019ef3646e
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>
Signed-off-by: Kevin Kane <kkane@microsoft.com>