ASACORE-2703 Initialize CNG before validating ECC public keys
Most of the ECC code assumes it's operating within the context of a
Crypto_ECC object, which makes sure the cached CNG provider handles
have been initialized. Adding key validation to ECCPublicKey breaks
this assumption.

This change moves the initialization code into a helper function that
is now called on both code paths.

Correct bug in Crypto subsystem shutdown code that was causing the
proper shutdown to never actually run.

Change-Id: Ia497d891a81c1dcb7c0609b8ed5de976f8648410
Signed-off-by: Kevin Kane <kkane@microsoft.com>