ASACORE-2710 Switch to MsgArg-based (de)serialization for Manifests
Rather than have a separate set of (de)serialization code for
manifests, use MsgArg-based serialization as is done elsewhere in the
code. This also makes things far more straightforward in the thin
client.

Remove the requirement that manifests being stored have to have the
thumbprint of our identity cert; if the security manager wants us to
store a manifest, we should do it. This is also consistent with the
thin client's behavior.

Fix bug in CertificateX509 to compute thumbprints over binary DER
encoding, instead of Base64-encoded PEM.

Change-Id: I8fcd46b328cff3e274ffd79802f193b7e63e8ac4
Signed-off-by: Kevin Kane <kkane@microsoft.com>