ASACORE-2722 Added placeholder SConscript for java ConfigClientSample
The services/config/java/samples/java folder has a SConscript that
references a non-existant SConscript file in its ConfigClientSample
sub-folder. This was intentional, as this parent SConscript has
the following comment:
   > The config client sample is an Android project and is not
   > currently set up to build using SCons.
   > Prevent Build Failure if None is returned.
However, in order to avoid the warning "Ignoring missing SConscript",
I have added this missing SConscript file in the ConfigClientSample
folder. This new file is simply a placeholder (returns jars=[]).
In future, when ConfigClientSample is setup to build using SCons,
then this placeholder SConscript file can be implemented to return
the actual jar.

Change-Id: Iaae988198712132d6fea63dc015a648ce6288714
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>