ASACORE-2661 ASACORE-2664 - Fix multiple sharing issues in KeyStore
With this change, multiple KeyStores that use the same persistent file can now
be shared reliably as access to the file is serialized. On Windows, access to
KeyStore file is also serialized between processes, allowing the file to be
shared across them. On platforms other than Windows, serialization is limited
to within the process (additional work on non-Windows platform is to be
addressed by ASACORE-2587).

Change-Id: Iacaff3f1c011b199ea998952ec3d21b8cac879af
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>
Signed-off-by: Dan Mihai <Daniel.Mihai@microsoft.com>