Mark ECDHE_PSK as deprecated
Mark the DefaultECDHEAuthListener constructor that takes a PSK as input
deprecated.  Remove all uses of this constructor from test code, replacing
it with the no-param constructor followed by a call to SetPSK.
Mark DefaultECDHEAuthListener::SetPSK deprecated, and update test and
sample code to suppress warnings.  Opened JIRA tasks to track work
of updating samples (in 16.10) and removing tests (in a future release).
Updated documentation to reflect PSK deprecation.

Add a null-check so that the util RandomString does not call
qcc::String(nullptr), causing an assertion to fail.

Change-Id: I3ba8be3d0cf9836322d8fd945688175831d1a3f7
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>