ASACORE-2749 Selective manifest sending
Only send manifests when there's a need, based on the message currently being sent.
Add on-demand triggering of manifest sending from from method calls and signals.

Remove PeerManifestState from PeerState and just store Manifest objects directly,
as there ended up being no need for it.

Disable one test broken by this behavior; JIRA ticket is opened to investigate whether
or not this is a scenario we actually need to support, as it's not clear.

Change-Id: Iaa298ea6943fffca3be35ed0dded7a69673781b7
Signed-off-by: Kevin Kane <kkane@microsoft.com>