ASACORE-2757 KeyStore::Clear() does not guarantee clearing of keystore
This fix ensures that another instance of KeyStore (sharing the same store)
wouldn't be able to resurrect cleared keys when KeyStore::Clear() is called
by the other instance.

Turns out the merge code is already doing the right thing by taking into
account the in-memory revision against the persistent revision, so all this
change does here is to not roll the in-memory revision number back to 0, so
the Clear operation would not be invalidated.

The new unit test "KeyStoreTest.keystore_clear" would fail before the change
and would succeed after.

Change-Id: Iede85374c93d91446b9ad47218a4fe54e6c89a44
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>