ASACORE-2818 ajtest's SecurityClaimApplicationTest.* fail every once in a while
Fix test crash due to a race condition in SecurityClaimApplicationTest during
shut down. The race occurs when the test body's stack frame has been torn down,
causing instance of applicationStateListener to be deleted before the security
manager could complete all of its callback.

In my investigation on a Linux failure (using the generated Core file), the BA
was still
handling "State" signal and the attempt to notify the listener caused a crash
since the listener was created on the stack and went down with the test body.

The fix here is to unregister the applicationStateListener(s) before exiting
the test body.

Change-Id: I581e7c6d855d2d0aed84e9d664fd74aa2f6e751c
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>