ASACORE-2772 Language-aware accessor functions needed for descriptions
Added setters and getters with a "languageTag" parameter for: interface description, interface property description, interface member description, interface member argument description.

The accessors store descriptions as "org.alljoyn.Bus.DocString" annotations with suitable language tags added. This means that calling any of the new setters is equivalent to calling one of the AddAnnotation() functions (for members, properties, etc.) with an "org.alljoyn.Bus.DocString" annotation name. Similarly, calling one of the getters is equivalent to calling one of the GetAnnotations() methods.
This single, annotation-based storage is then used by the Introspect() method.

These accessors will replace the old ones, which are not language-aware and work in conjunction with optional translators to get descriptions in multiple languages. The old methods, as well as the translator-related functions, are to be deprecated under ASACORE-2748. Currently, both the old and new setters can be used to set descriptions which will then be available when introspecting, but it is recommended that only one of the methods: the old or the new one, is used on a single InterfaceDescription. If there are descriptions set by both the new and the old (legacy) method in a single InterfaceDescription, legacy descriptions will be preferred by the introspection methods.

Change-Id: Ic573a23200849142f530b848f379ab2796fd88c2
Signed-off-by: Karol Matuszkiewicz <kmatuszkiewicz@affinegy.com>