ASACORE-2818 ajtest's SecurityClaimApplicationTest.* fail every once in a while
Fix a test crash due to a race condition in SecurityClaimApplicationTest
during shut down. The race occurs when the test stack frame has been torn down,
causing the instance of applicationStateListener to be deleted before the
security manager could complete its callback.

In my investigation on a recent Jenkins failure (using a Linux generated Core
file), security manager's bus attachment was still in the middle of handling
"State" signal, which came in *after* the test has torn down the listener and
was processing the test's TearDown method.

The fix here is to unregister the applicationStateListener(s) before exiting
the test body.

Change-Id: I581e7c6d855d2d0aed84e9d664fd74aa2f6e751c
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>