ASACORE-2910 KeyStoreTest.basic_store_load failing on Android
Was failing because FileSink("keystore_test") was not creating
the file. This turned out to be because it needed the full path
of the file to be specified. Now we specify the path using
qcc::GetHomeDir(). Also, added ASSERT to verify that the
resulting FileSink instance IsValid (i.e. was created).

Change-Id: Iac76ea591d501fb5f76d55a65efd12b5d1889290
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>