ASACORE-2916 KeyStoreTest.basic_store_load doesn't create keystore in HOME dir
The basic_store_load test was attempting to create the keystore_test
file in the current directory (which it may not have permissions to
do). In order to be like the other tests, the file will now be created
in the HOME directory, determined using the qcc::GetHomeDir() function.
Also added ASSERT to verify that the file is created.

Change-Id: I1b2a3112fc42c18d0f0d47c8436d6b6e4101797b
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>