ASACORE-2916 KeyStoreTest.basic_store_load doesn't create keystore in HOME dir
The basic_store_load test was attempting to create the keystore_test
file in the current directory (which it may not have permissions to do).
In order to be like the other tests, the file will now be created in the
HOME directory, determined using the qcc::GetHomeDir() function.
Also added ASSERT to verify that the file is created.
Note that we are using qcc::String with this change due to the fact
that qcc::FileExists, qcc::FileSink, and qcc:GetHomeDir()
all opearte on qcc::String. So using qcc::String in the test would
avoid implicit constructor calls/conversions.

Change-Id: Id67c705f445bb90711ba711ddb3212baee4fb5d0
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>