ASACORE-2912 Expand PermissionConfigurator to sign manifests
PermissionConfigurator is used to issue certificates using a
BusAttachment's key pair, which is the desired way to implement
some delegation scenarios. The methods on the Manifest class to
sign manifests require an ECCPrivateKey, but there is no way in
the public API to acquire a bus attachment's private key in an
ECCPrivateKey object.

This change adds a SignManifest method for signing given a
thumbprint, and a ComputeThumbprintAndSignManifest method to do
so given a certificate, which wrap the equivalent functionality
provided by the Manifest class, using the private key owned by the
BusAttachment associated with the PermissionConfigurator.

Change-Id: Ib95b7277bb242db489987b666f06f2d12d60b3c2
Signed-off-by: Kevin Kane <kkane@microsoft.com>