ASACORE-2841 channel names with periods
Channel names can now have multiple periods. These channel names will
show up in the list of available channel names and there will be no
alljoyn exception when trying to connect with these channels.

Added check for string length. If less than or equal to prefix + 1 (1
for the period), then continue. This would apply to an empty channel
name or a wellknown name that doesn't have the same prefix.

Change-Id: If08b9ba906669360ed4a40d986e2e9eeed2fc0cb
Signed-off-by: George Tang <gtang@affinegy.com>