ASACORE-1454 Offline Security 2.0 C/C++ claim/management APIs
Refactor methods in ClaimableApplication and ManagedApplication interfaces
to allow local versions to be added to PermissionConfigurator, for provisioning
Security 2.0 credentials and policy out-of-band, allowing provisioning before
an app connects to the bus. The new APIs for PermissionConfigurator mostly match
the signatures of equivalent methods in SecurityApplicationProxy used for
online claiming and management.

Add additional APIs and unit tests for C API for PermissionConfigurator.

Change-Id: I0812818c35d6215d1e079ca64d0ebaec35954070
Signed-off-by: Kevin Kane <kkane@microsoft.com>