ASACORE-2856 Remove more uses of memcpy on ECC crypto objects
Change uses of memcpy in KeyInfoNISTP256 to copy an ECCPublicKey and
an ECCSignature to use the assignment operator, to remove
the dependency of this code on any future changes to the internal
structure of these objects.

Deprecate GetPublicCtx and SetPublicCtx methods in KeyInfoNISTP256.
The only purpose of these appeared to be to implement the copy
constructor and assignment operator, which now have been changed to
use the assignment operators of the objects being copied.

Change-Id: I68fe41d700e6aa9dcb5491110bf16e5348012774
Signed-off-by: Kevin Kane <kkane@microsoft.com>