ASACORE-2774: Develop Sample Apps for KeyStoreListener
Fixed KeyStoreListener (KSL) interface to not use in API reference
to non public KeyStore class but to use void* context wrapper instead.
Non virtual PutKeys() and GetKeys() use this context internally.

Changed Darwin fork of KSL to symlink as it was always the same as Posix,
but after changes for ASACORE-2587 in Posix both versions diverged
unintentionally.

Change-Id: Ie91449eac6d7571e50b8d99295d5b6f35a23e141
Signed-off-by: Krystian Zlomek <kzlomek@affinegy.com>