ASACORE-2924 return custom error messages
The code now throws custom error messages in the exception objects on
get and set property. The errors were prefixed with ER_FAIL, because of the
ErrorMsg function in Message_Gen. So I removed some code from
Message_Gen. I have not yet configured this to work for
getAllProperties. I wanted to upload to get feedback on syntax or
comments to improve the code here.

Change-Id: I5d69a2203a57fc93db46ddacbe8e94fa2e66eec7
Signed-off-by: George Tang <gtang@affinegy.com>