ASACORE-2924 return custom error messages
The code now throws custom error messages in the exception objects on
get and set property. There are a few problems that I still need to iron
out. One problem is the error messages are prefixed with ER_FAIL for
some reason. I have not yet configured this to work for
getAllProperties. I wanted to upload to get feedback on syntax or
comments to improve the code here.

Change-Id: I5d69a2203a57fc93db46ddacbe8e94fa2e66eec7
Signed-off-by: George Tang <gtang@affinegy.com>