ASACORE-3015 Export named constants from C API on Windows
CLAIM_CAPABILITIES_DEFAULT and PERMISSION_MANAGEMENT_SESSION_PORT are
mentioned in the C API header files for use by callers, but they are
not actually exported as symbols from the C API on Windows. This
causes linker errors when the constants are used in application code.

Instead, replace these constants with functions that will provide the
constants.

Change-Id: I59677a6fbe4727f3e34587a02ec4eb17781bb5c8
Signed-off-by: Kevin Kane <kkane@microsoft.com>