ASACORE-1454 Support external signatures of certificates/manifests
Add APIs to support performing signatures on certificates and
manifests outside of the AllJoyn code.

Add C++ APIs to CertificateX509 to retrieve the "To Be Signed" (TBS)
portion of the structure, over which the digest and signature are
computed. Add a method to set a signature to the object.

Add C and C++ APIs to support retrieving the SHA-256 digest of a
manifest, and to set a signature on the object.

Change-Id: I78913589136c71526cc9f5512f160bd2e208eee8
Signed-off-by: Kevin Kane <kkane@microsoft.com>