ASACORE-3044 ASACORE-3045 Fix Security 2.0 XML schema doc problems
Currently committed versions of XSDs for policy, manifest, and
manifest templates use features from XML 1.1. To provide XSDs that
are compatible with XML 1.0 parsers, as is already done for the
introspection XSD, rename the XML 1.1 versions of these schemas to
have "-1.1" in the filename, and modify the existing schemas to be
XML 1.0-compliant.

The "-1.1" version of these files are exact copies of what is
currently committed, and are not new content.

Correct the regular expression for public key PEM strings in the
policy XSD to accept valid PEMs; current regex does not include
forward slashes, plus signs, and newlines, all of which are valid
in such a string.

Change-Id: Iabbe519374bb9f1fbc91337805c1e214c7ec640c
Signed-off-by: Kevin Kane <kkane@microsoft.com>