ASACORE-2730 Fix IpNameServiceImpl::PurgeAndUpdatePacket infinite loop
IpNameService could spin forever on some mDNS queries that contain one or more
search criteria delimiter (";") in the TEXT record. For example, when a client
needs to send the following query:

    n_1="org.alljoyn.sl.*"
    ;_2=""
    txtvers="0"

And here's what PurgeAndUpdatePacket does:

  (A) uint32_t numSearch = searchRData->GetNumSearchCriteria();
  (B) for (uint32_t k = 0; k < numSearch; k++) {
  (C)    String crit = searchRData->GetSearchCriterion(k);
  (D)    if (std::find(set_union_tcp_udp.begin(), set_union_tcp_udp.end(),
                       crit) == set_union_tcp_udp.end()) {
  (E)        searchRData->RemoveSearchCriterion(k);
  (F)        k--;
  (G)        numSearch = searchRData->GetNumSearchCriteria();
  (H)    }
  (I) }

1. The function first gets the number of search criteria by calling
GetNumSearchCriteria (A) and this function simply counts number of concatenated
field separated by semicolon.

2. Then the Purge function attempts to retrieve the search criterion at index
k (C), which could fail with an empty string because the search function also
does some field filtering (txtvers field, for instance, is never included in a
search criterion).

3. Since an empty search criterion is not in the query we're interested, the
check at line (D) succeeds and we're going to try to attempt to remove.

4. The function remove function, RemoveSearchCriterion (E), fails since we're
attempting to remove a bogus search criterion. Since (E) effectively is a
no-op, the index decrement at (F) now causes the loop to never exit.

The fix here is to check whether the remove operation in #4 is successful
before decrementing the criterion index (F).

Change-Id: I88cd242d16d44ca8b1d23189e402375315559d96
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>