ASACORE-3133: Fixing security manager crashes.
Fixed the security manager and the sample C++ app
to prevent the crashes. In order to do that I've
updated the core code to handle (End|Start)Management
during "Claim" and "Reset" calls automatically.

This forced changes in some tests and the sample
security manager, due to the fact that they should
now use PermissionConfigurator to claim themselves
instead of SecurityApplicationProxy objects.

Change-Id: Ib12560505dca71d93a15fe51358e44b5a63fe3cf
Signed-off-by: Pawel Winogrodzki <pawelwi@microsoft.com>