ASACORE-2067 ASACORE-3141 Improve membership cert filtering
Fix filtering bug where membership certs issued by the same
security manager but using a different trust anchor were excluded.
If a membership cert is issued by one of the trust anchors, and the
peer's identity cert is issued by one of the trust anchors, but not
necessarily the same one, we can conclude the peer is in the same
security domain and so is trusted to receive certificates issued by
our security manager.

Otherwise, to support foreign certificate scenarios, only send if the
membership cert and the peer's issuer cert are issued by the same
authority.

Change-Id: Ie668fdb4b24abbd8dec64057617647bcacccec1c
Signed-off-by: Kevin Kane <kkane@microsoft.com>