ASACORE-3136 Add new API to permanently delete the key store
This change allows app to delete the key store based on the application name.
As part of this change, the previous (duplicate) function for doing the same
in C binding is removed.

With this update, if the file does not exist, something like this will be
shown in verbose log:

  ...s\windows\FileStream.cc:46 | DeleteFile(...\AppData\Local/
  .alljoyn_secure_keystore/keystoretest_shared_keystore) returned
  ERROR_FILE_NOT_FOUND, converting to ER_EOF

If the deletion fails, something like this is shown:

  ...s\windows\FileStream.cc:49 | DeleteFile(...\AppData\Local/
  .alljoyn_secure_keystore/keystoretest_shared_keystore) failed
  (0x5): ER_OS_ERROR

Change-Id: I3972d1b64dba5d51d97da1101699e22c7ba3a923
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>