ASACORE-3169 Document best practice for local claiming to avoid race
When claiming a bus attachment locally, there is a possibility that,
between the time the bus attachment is connected to the bus and peer
security is enabled, and when the local process calls the claim API,
a rogue security manager on the bus will call Claim over the network
and claim the bus attachment first, before the local Claim call is
made.

This change adds documentation to the C and C++ Claim calls in
PermissionConfigurator to show how the bus attachment can be properly
configured to reject any attempts over the network to claim it, and
thus guarantee the local Claim is successful.

Update Secure Door C/C++ samples to employ this best practice.

Change-Id: I0ee7e1238c193b4d981f873430a104d26550794c
Signed-off-by: Kevin Kane <kkane@microsoft.com>