ASACORE-3190 Implement SecurityApplication iface
I implemented the proxy side of the Security Application
interface and added a sample app that works with the
SecureDoor samples.

future work:
-more Security 2.0 ifaces, sample apps, and unit tests

-MANIFEST.MF is in alljoyn_java/.gitignore
However, builds will fail if this file is not present.
This leads me to believe it is automatically generated
or should be. However, when cloning a new core/alljoyn
repo, these files exist in the alljoyn_java/samples folders.
The confusing part to me is how are these files untracked
and yet in the git repo.

Change-Id: I293af5f7d5da4e9e8fa2f74552d73a8e8e7c8833
Signed-off-by: George Tang <gtang@affinegy.com>