tASACORE-3055 split alljoyn_java
In order to implement the permissionconfigurator in a separate
c++ file, it must include a header for JBusAttachment. JBusAttachment
depends on
JAboutObject
JAuthListener
JKeyStoreListener
JSignalHandler
PendingAsyncJoin
PendingAsyncPing

It is not possible to include the c++ file alljoyn_java from the
PermissionConfigurator, because including it will redefine several
classes. So JBusAttachment needed to be in a separate header file. The
same reasoning was used to place the definition of the above classes in
separate header files.

Change-Id: I0241a087f827cf2e9db6b100fe9cddbd8c1725d9
Signed-off-by: George Tang <gtang@affinegy.com>