ASACORE-3118 Deprecate old description accessors in Java bindings
Deprecating the old description accessors in InterfaceDescription
in favor of using the new language-aware accessors.

Also deprecating the BusInterface java annotation
@BusInterface(descriptionTranslator) as it pertains to setting
a description translator on InterfaceDescription (which is
itself being deprecated).

Note that there are other java annotations that currently
will result in the old description accessors being called,
but they are not deprecated in this commit. These include:
@BusInterface(description, descriptionLanguage)
@BusMethod(description)
@BusSignal(description)
@BusPropety(description)
as they can be repurposed in the future to call the new
language-aware accessors in order to conveniently set
a description if only one langauge is to be used in
a bus inteface's descriptions.

Change-Id: I262636ebf1b01f65cadf8c5fc52e27921cfda21a
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>