ASACORE-3108: Android add support for arm64 architecture
Updated scons scripts to allow build of Android alljoyn_java
for arm64 and x86_64 architectures (i.e. 64 bit). Requires
minimum NDK Level of API 21 (Lollipop; Android 5.0).

The alljoyn_java android java samples will also be built
in arm64 and x86_64 as part of the scons android build
when specifying a CPU in the scons build of either
arm64 or x86_64.

Also, the android jni samples (cpp) can now also be built
for other architectures (e.g. arm64, x86_64) where
previously they had been fixed to only build as arm.
As usual, this is done after building the alljoyn_java.so,
but now a custom ABI value can be passed into the
ndk-build command to specify which architecture to
use (arm64, x86_64, etc), where previusly was always
fixed as arm. Note that if this ABI value is not specified
for the ndk-built tool, then it defaults to arm.

Change-Id: I074a8294ed69d5dbc6922eaa48d32e5642fc7d8a
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>