ASACORE-3220: Update InterfaceDescription(java) to support dynamic interfaces
Support creation of an InterfaceDescription instance based on a dynamic
bus object and its dynamic interface defs. Existing InterfaceDescription
implementation now supports add/configure of properties, methods, and
signals based on a dynamic interface definition. Assignment of java
Method member callbacks is determined via the DynamicBusObject API.

Changed the InterfaceDescription class to be public so that dynamic
interfaces can be created from outside of the org.alljoyn.bus package.

Many of the new methods in InterfaceDescription are based on existing
methods that built the interface using class Annotations. These new
version of these methods provide a means to build the interface using
the new dynamic InterfaceDef (and related) definition classes.

Change-Id: Ifd09f88cc3268d56d3fff5404291c8a1df855be6
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>