ASACORE-3279 Harmonize and improve Crypto_ECC across platforms
Make the generic ECC implementation make sure a DH or DSA public or
private key has been set before returning it, to prevent applications
from using key material before generating or loading it. nullptr is
returned and an assertion is raised in debug build if this happens,
as this is a caller bug.

Improve the tracing on ECCPublicKey::Import for both generic and CNG
to specifically state the imported public key is invalid as the
reason for failure (the point does not lie on the elliptic curve).

Update documentation for Crypto_ECC to accurately describe this
behavior.

Change-Id: I3a003f62251c0700cd83eb6d4e0154b8004dc949
Signed-off-by: Kevin Kane <kkane@microsoft.com>