ASACORE-3156 Sec2.0 GetPermissionConfigurator
BusAttachment has a function GetPermissionConfigurator
It looks like PermissionConfigurator is managed by the BusAttachment
and is constructed when a BusAttachment::Internal is constructed.
Thus I removed exposure of the PermissionConfigurator constructor in
java. And I also removed its destroy function, since it appears the bus
should manage the memory of the PermissionConfigurator.

Change-Id: I539db92169b4ac6d3e86b3be8748acafeb8acef3
Signed-off-by: George Tang <gtang@affinegy.com>