ASACORE-3156 Sec2 PermissionConfigurationListener
added jni and java implementation of PermissionConfigurationListener
added to the test for PermissionConfigurator
added enablePeerSecurity exposure that uses the
PermissionConfigurationListener
Fixed some problems with GetSigningPublicKey
 -method signature
 -return type
It appears KeyInfoECC in common doesn't return a valid public key
on GetPublicKey, so you must use KeyInfoNISTP256 instead.

Also, java will have a class cast exception if the java method
definition returns KeyInfoECC and if the returned object is
casted to KeyInfoNISTP256, even if you are calling KeyInfoNISTP256
constructor from the JNI layer.

Change-Id: I826642b5df5a1aaf3bc62117986acbac5c5bda65
Signed-off-by: George Tang <gtang@affinegy.com>