ASACORE-3295 ObserverTest.Multi test failing in Jenkins
I was able to reproduce ObserverTest.Multi test failure in debug
mode with App Verifier enabled. The failure starts off like this:

    ...\ObserverTest.cc(827): error: Value of: WaitForAll(events)

Turns out with AppVerif enabled, the test sometimes does not have
sufficient time to complete as indicated by sessionless messages
were still waiting to be delivered:

    0:039> ??bundledRouter->ajBusController->sessionlessObj.
        timer.timerImpl->alarms._Mypair._Myval2._Myval2._Mysize
    >7
    // There were 7 About announcements waiting to be picked up

During this time SessionlessObj only had 1 timer thread, and was
busy servicing other alarms. In my repro, the controller thread
was blocked on heap allocation by the _LocalEndpoint dispatcher,
so there was no deadlock, just insufficient time to complete the
test.

I also looked at a handful of other instances of this failure and
found that everything worked as advertised and that all the failures
were due to the the test's timeout. In these repros, I noticed a
significant increase in heap contentions when AppVerif is enabled.
It appears Observer tests require massive thread interactions as
well as small buffer allocations.

Because of the above findings, the solution I chose for fixing this
is to just increase the test's timeout. I verified that this fix
is sufficient by repeating the test 100 times (--gtest_repeat=100)
all with AppVerif enabled. Previously before the fix, the test
would just fail consistently when repeating just < 10 times).

Change-Id: Idfea657910521bc155e6db207fca22809d53955d
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>