ASACORE-3365 CertificateX509 unit tests
copied unit tests from common CertificateECC test

1 line fails from these unit tests in CertificateX509Test line 763 so I
commented it out.

Also:
added getType to CertificateX509.
made constructors for ECCPublicKey and ECCPrivateKey public.
added setters and getters for the data types that they surround.
added equals function to ECCPrivateKey
fixed a bug in the method sign of CertificateX509 that was causing
crashes.
changed the exceptions of CryptoECC to BusExceptions

Change-Id: I579e3b6dae4ea4300e328451f043cc701ef8c1be
Signed-off-by: George Tang <gtang@affinegy.com>