ASACORE-3353 Codegen'ed apps cannot interop with non-codegen'ed apps
The AllJoyn Unified Introspection XML, introduced in 16.04, has a bug
where the generated list of annotation nodes (under an interface, for
instance) are always sorted. This effectively can lead to
incompatiblities between apps, particularly ones that utilize these
out-of-order annotations such as codegens and bridging translators.

With this fix, we will propose to include the following into the
release note of the upcoming AllJoyn release:

"An application that introspects and finds multiple annotations based
on AllJoyn 16.04 should treat these annotations as unusable. To
determine the AllJoyn framework version, an application can query the
About data."

Additionally it is recommended that the existing codegens, such as
Microsoft AllJoyn Studio, be patched to look for AllJoyn version when
it dynamically introspects an interface via About. The code generated
from a 16.04 interface shall omit names that come from annotations,
or alternatively rename them to Argument1, Argument2, for example.

Change-Id: Ib5f717f09e4a514b4984ee2f4e7953a93f030a2b
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>