ASACORE-2165: privacy vulnerability in security 2.0 Reset() in Java
The vulnerability (i.e. reset() needs to allow the app to delete
its persistent state via a callback registered by the app),
was resolved as part of an earlier commit under another JIRA
issue. However, this latest commit updates unit tests in
order to add coverage for this case case.

Change-Id: Ice052f3d6f622c5876fd8ea13c1dc7228141ef1a
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>