ASACORE-3200 Thread safety improved
In order to to remove various possible data race situations I added:
- AtomicFetch(..) for all supported platforms
- atomic class wrapper for int32_t, uint32_t and bool

I added the thread sanitizer logs to the gerit change set (as simple comments).

The new atomic wrapper classes are just a substitute for std::atomic<>
as this module is not yet available (gcc 4.6).
However I designed qcc::atomic in a way that it is easily exchangeable when
std::atomic API becomes available.

The new functions AtomicFetch and AtomicSet guarantee a thread safe access
to critical variables (read and write). Unfortunately I was not able to carefully
test Windows and Mac variants. Hope that somebody else can do this?

Change-Id: I29bab5d5c2617bc75d4c236526c2fdb1675d1571
Signed-off-by: Torsten May <t.may@technisat.de>