ASACORE-3390: Junit test is failing on Windows and Android
Fixed issue in the CryptoECCTest test. Not calling sequence
of operations in the correct order.

Fixed issue in SecurityManagementTest where the application
state can unexpectedly be CLAIMED instead of CLAIMABLE.
Also saw this in linux junit and seemed to be associated
with keystore listener not set.

Changed PermissionConfiguratorTest to register auth listener
with ECDHE_NULL, with keystore file, and permission config
listener.

Change-Id: Ibbe547ee951ab40565c3335da65bce3953c2eb6b
Signed-off-by: Paul Sigurdson <psigurdson@affinegy.com>