ASACORE-3467 workaround v-table corruption when same obj name is used
The Microsoft Visual Studio compilers/linkers can generate wrong code
if there are multiple class implementations of the same name deriving
from the same base class (even when they are in different source
files). The consequence seen in this bug is a memory corruption in
the v-table of an unrelated object.

The "fix" done here is to workaround this complier bug by scoping
each of these duplicate implementations in its own namespace.

Change-Id: I976b5ac206a81da3f78c9a7b69b5dbeb7d5fabbc
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>