ASACORE-3467 Fix C++ violations as they can cause v-table to corrupt
This change fixes One Definition Rule violations in AllJoyn. This is
defined in the ISO C++ Standard (ISO/IEC 14882) 2003, section 3.2.

The Microsoft Visual Studio linkers can generate wrong code if there
are multiple class implementations of the same name deriving from the
same base class (even when they are in different source files). The
consequence seen in this bug is a stack corruption, causing the
v-table of an unrelated object to corrupt.

The "fix" done here is to scope the conflicting implementations into
their own (anonymous namespace).

Change-Id: I976b5ac206a81da3f78c9a7b69b5dbeb7d5fabbc
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>