ASACORE-3472 PermissionConfigurator Statement Test
There is a typo in the JNI code. This confirms the need for a 16.10a
release.

Wrote unit tests that covers all functions of PermissionConfigurator
except getConnectedPeerPublicKey, since this function is used with
CredentialAccessor and since CredentialAccessor is not available in the
Java API. Credential Accessor extracts the bus guid and is used as a
parameter for the function getConnectedPeerPublicKey.

Change-Id: Ie23d80e3864538413d63efb16104888270a556d9
Signed-off-by: George Tang <gtang@affinegy.com>